Refactor time semantic ops mess a bit
