+++ simag_core/src/agent/lang/cls_decl.rs
@@ -1,10 +1,12 @@
 use std::collections::HashMap;
+use std::convert::TryFrom;
 use std::sync::{atomic::AtomicBool, Arc};
 
 use super::{
     common::*,
     logsent::{LogSentResolution, ParseContext},
     parser::ClassDeclBorrowed,
+    time_semantics::{TimeArg, TimeOps},
     *,
 };
 use crate::agent::kb::{bms::BmsWrapper, repr::Representation, VarAssignment};
@@ -91,6 +93,7 @@ impl<'a> ClassDecl {
         for arg in self.op_args.as_ref().unwrap() {
             match *arg {
                 OpArg::TimeDecl(_) | OpArg::TimeVarAssign(_) => {
+                    let arg = TimeArg::try_from(arg).unwrap();
                     v = Some(arg.get_time_payload(assignments, value));
                 }
                 OpArg::OverWrite => {
@@ -107,54 +110,6 @@ impl<'a> ClassDecl {
             bms.new_record(None, value, None);
             bms
         }
-    }
-
-    pub(in crate::agent::lang) fn get_time_decl(&self, var0: &Var) -> bool {
-        if self.op_args.is_none() {
-            return false;
-        }
-        for arg in self.op_args.as_ref().unwrap() {
-            if let OpArg::TimeVarFrom(ref var1) = *arg {
-                return var1.as_ref() == var0;
-            }
-        }
-        false
-    }
-
-    pub(in crate::agent::lang) fn get_times(
-        &self,
-        agent: &Representation,
-        var_assign: Option<&HashMap<&Var, &VarAssignment>>,
-    ) -> Option<Arc<BmsWrapper>> {
-        let arg = &self.args[0];
-        match *arg {
-            Predicate::FreeClsMemb(ref free) => {
-                var_assign?;
-                if let Some(entity) = var_assign.as_ref().unwrap().get(&*free.term) {
-                    if let Some(grounded) = entity.get_class(free.parent.get_name()) {
-                        if free.grounded_eq(grounded) {
-                            return grounded.bms.clone();
-                        }
-                    } else {
-                        return None;
-                    }
-                } else {
-                    return None;
-                }
-            }
-            Predicate::GroundedMemb(ref compare) => {
-                let entity = agent.get_obj_from_class(self.get_name(), &compare.term);
-                if let Some(grounded) = entity {
-                    if *grounded == *compare {
-                        return grounded.bms.clone();
-                    }
-                } else {
-                    return None;
-                }
-            }
-            _ => return None, // this path won't be taken in any program
-        }
-        None
     }
 
     pub fn get_time_payload(&self, value: Option<f32>) -> Option<BmsWrapper> {
@@ -220,6 +175,50 @@ impl<'a> ClassDecl {
             }
         }
         Ok(())
+    }
+}
+
+impl OpArgsOps for ClassDecl {
+    fn get_op_args(&self) -> Option<&[common::OpArg]> {
+        self.op_args.as_ref().map(|r| r.as_slice())
+    }
+}
+
+impl TimeOps for ClassDecl {
+    fn get_times(
+        &self,
+        agent: &Representation,
+        var_assign: Option<&HashMap<&Var, &VarAssignment>>,
+    ) -> Option<Arc<BmsWrapper>> {
+        let arg = &self.args[0];
+        match *arg {
+            Predicate::FreeClsMemb(ref free) => {
+                var_assign?;
+                if let Some(entity) = var_assign.as_ref().unwrap().get(&*free.term) {
+                    if let Some(grounded) = entity.get_class(free.parent.get_name()) {
+                        if free.grounded_eq(grounded) {
+                            return grounded.bms.clone();
+                        }
+                    } else {
+                        return None;
+                    }
+                } else {
+                    return None;
+                }
+            }
+            Predicate::GroundedMemb(ref compare) => {
+                let entity = agent.get_obj_from_class(self.get_name(), &compare.term);
+                if let Some(grounded) = entity {
+                    if *grounded == *compare {
+                        return grounded.bms.clone();
+                    }
+                } else {
+                    return None;
+                }
+            }
+            _ => return None, // this path won't be taken in any program
+        }
+        None
     }
 }
