+++ simag_core/src/agent/lang/parser.rs
@@ -1,7 +1,7 @@
 //! # Grammar for the SIMAG declarative language
 //! ```BNF
-//! @comments			::	regex: \/\*(.*)\*\/ (multiline)
-//! @eol_comments 		::	/#.*?$/
+//! @comments           ::  regex: \/\*(.*)\*\/ (multiline)
+//! @eol_comments       ::  /#.*?$/
 //!
 //! scope = '(' ${var_decl}* ${skol_decl}*
 //!             (class_decl | func_decl | scope | multiple)
@@ -12,19 +12,19 @@
 //! skol_decl = '(' 'exists ' ${term [':' op_arg]}','+ ')' ;
 //! class_decl = term ['(' ${op_arg}','+ ')'] args ;
 //! func_decl = 'fn::' term ['(' ${op_arg}','+ ')'] args
-//! 		  | 'fn::' term '(' ${op_arg}','+ ')' ;
+//!             | 'fn::' term '(' ${op_arg}','+ ')' ;
 //! args = '[' ${ arg }';'+ ']';
 //! arg = term [',' uval] ;
 //! uval = 'u' comp_op number ;
 //! op_arg = (string|term) [comp_op (string|term)] ;
-//! icond_op    =	':=' ;
-//! and_op      =	'&&' ;
-//! or_op		=	'||' ;
-//! logic_op	=    '<=>'
-//! 		    |    '=>'
+//! icond_op    =   ':=' ;
+//! and_op      =   '&&' ;
+//! or_op       =   '||' ;
+//! logic_op    =    '<=>'
+//!             |    '=>'
 //!             |    or_op
-//!             |	 and_op ;
-//! comp_op	= ('=' | '<' | '>' | '>=' | '<=' ) ;
+//!             |    and_op ;
+//! comp_op = ('=' | '<' | '>' | '>=' | '<=' ) ;
 //! term = regex: \$?[a-zA-Z0-9_]+ ;
 //! number = regex: -?[0-9\.]+ ;
 //! string = regex: ".*?"|'.*?' ;
