+++ simag_core/src/agent/lang/time_semantics/time_ops.rs
@@ -1,8 +1,10 @@
 use std::collections::HashMap;
+use std::convert::TryFrom;
 use std::sync::{atomic::AtomicBool, Arc};
 
+use super::TimeArg;
 use crate::agent::kb::{bms::BmsWrapper, repr::Representation, VarAssignment};
-use crate::agent::lang::{OpArg, OpArgsOps, Var};
+use crate::agent::lang::{common::OpArg, OpArgsOps, Var};
 
 pub(in crate::agent) trait TimeOps: OpArgsOps {
     fn get_own_time_data(
@@ -20,6 +22,7 @@ pub(in crate::agent) trait TimeOps: OpArgsOps {
         for arg in self.get_op_args().unwrap() {
             match *arg {
                 OpArg::TimeDecl(_) | OpArg::TimeVarAssign(_) => {
+                    let arg = TimeArg::try_from(arg).unwrap();
                     v = Some(arg.get_time_payload(assignments, value));
                 }
                 OpArg::OverWrite => {
@@ -45,5 +48,16 @@ pub(in crate::agent) trait TimeOps: OpArgsOps {
         var_assign: Option<&HashMap<&Var, &VarAssignment>>,
     ) -> Option<Arc<BmsWrapper>>;
 
-    fn get_time_decl(&self, var0: &Var) -> bool;
+    fn get_time_decl(&self, var0: &Var) -> bool {
+        if let Some(args) = self.get_op_args() {
+            for arg in args {
+                if let OpArg::TimeVarFrom(ref var1) = *arg {
+                    return var1.as_ref() == var0;
+                }
+            }
+            false
+        } else {
+            false
+        }
+    }
 }
