--- simag_core/src/agent/time_semantics/mod.rs
@@ -1,1 +1,0 @@
-mod time_ops;
